backend/requirements.txt
fastapi
sqlmodel
uvicorn
pyjwt
pydantic
python-dotenv
requests
cachetools
pytest==7.0.1
passlib[argon2]
pydantic[email]
